+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This is used for recording Invader's changes. This changelog is based on
 - New tool: invader-convert - Converts between various tag types (gbxmodel to
   model, model to gbxmodel, shader_transparent_chicago_extended to
   shader_transparent_chicago)
+- invader-archive: Added `-e` and `-E` which check if a tag exists in a
+  a specified directory (multiple directories can be specified). Using `-E`
+  instead of `-e` does a functional comparison, too.
 - invader-build: Added `-g mcc-retail` which targets retail MCC (intended for
   CEA).
 - invader-build: Added `-g xbox-tw` and `-g xbox-jp` which targets the Taiwanese